api/requirements.txt
fastapi
uvicorn
motor
mysql-connector-python